refactor(vmm): move ui progress helpers into mvm-vmm::host
